added doc comments and changed name of bucket class
